Alteração do idFormaDoc de Integer para Long

+++ sigaex/src/main/java/br/gov/jfrj/siga/vraptor/ExModeloController.java
@@ -122,7 +122,7 @@ public class ExModeloController extends ExSelecionavelController {
 			final String tipoModelo, final String conteudo,
 			final ExClassificacaoSelecao classificacaoSel,
 			final ExClassificacaoSelecao classificacaoCriacaoViasSel,
-			final String descricao, final Integer forma, final Long nivel,
+			final String descricao, final Long forma, final Long nivel,
 			final String arquivo, final String diretorio, final String uuid,
 			final Integer postback) throws Exception {
 		assertAcesso(VERIFICADOR_ACESSO);
